Add files via upload
Total time exceeded 4 hours on account of my Unity not being setup to produce Unity webGL projects.  After a lengthy troubleshooting process, I was finally able to compile a quick take on the project.  

Unity projects are made up of MonoBehaviours (C# classes) that are managed by the Unity Engine.  These behaviours can be attached to anything exposed by the unity engine (Sprites, physics scripts, UI elements, etc).  

The hangman project consists of a model layer and a presentation layer.  The model layer manages which letters have been guessed, and the word, clue, and theme, as well as simple persistence, which is mostly managed by Unity. 

The view displays the model information via binding by Anonymous Action.  When values in the model are changed, they update the front end by having the front end first register appropriate actions with the model.  This allows the model to be decoupled from the view like some kind of data binding.