Add files via upload
The Build folder contains the built binary and index.html that needs to be hosted for the project to be demoed.  This has worked on Chrome and Edge on my machine, but only after it was built directly from Unity, which runs it's own server to host the WebGL project.

I am hoping very much that you can get this to run!  I have read repeatedly that most browsers cannot run WebGL code from a local Uri due to some new security policies, so it does need to be hosted somewhere as far as I can tell.